+++ libft_og/ft_strlcat.c
@@ -6,15 +6,17 @@
 /*   By: aldiaz-u <aldiaz-u@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 17:18:42 by aldiaz-u          #+#    #+#             */
-/*   Updated: 2025/04/08 18:06:00 by aldiaz-u         ###   ########.fr       */
+/*   Updated: 2025/04/09 17:37:19 by aldiaz-u         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+#include "stdlib.h"
+
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	dest_size;
-	unsigned int	src_size;
-	unsigned int	i;
+	size_t	dest_size;
+	size_t	src_size;
+	size_t	i;
 
 	dest_size = 0;
 	src_size = 0;
